refactor: move vendored jsonpath into its own module
Per review feedback, keep the vendored Goessner jsonpath engine in a
separate file (jsonpath_vendor.py) so it stays isolated from code we
maintain ourselves. envoy_reader now imports it with
`from .jsonpath_vendor import jsonpath`. No behaviour change.
